--- a/intel_extension_for_transformers/llm/runtime/graph/models/mpt/mpt_utils.cpp
+++ b/intel_extension_for_transformers/llm/runtime/graph/models/mpt/mpt_utils.cpp
@@ -48,6 +48,7 @@ void model_load_internal(const std::string& fname, model_archs arch, model_conte
   ms->init(fname.c_str(), lctx, n_ctx, n_gpu_layers, use_mmap, use_mlock, vocab_only);
   ms->load(lctx, progress_callback, progress_callback_user_data);
 
+  lctx.support_jblas_kv = true;
   lctx.t_load_us = ne_time_us() - lctx.t_start_us;
 }